+++ test-common/src/main/java/com/layer/xdk/test/common/stub/MessagePartStub.java
@@ -8,6 +8,7 @@ import com.layer.sdk.listeners.LayerProgressListener;
 import com.layer.sdk.messaging.Message;
 import com.layer.sdk.messaging.MessagePart;
 
+import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.util.Date;
 import java.util.UUID;
@@ -42,7 +43,7 @@ public class MessagePartStub extends MessagePart {
 
     @Override
     public InputStream getDataStream() {
-        return null;
+        return new ByteArrayInputStream(mData);
     }
 
     @Override