+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
   * Adding a mutator for the `MessageModelManger` singleton to the `ServiceLocator` (AND-1474)
+  * Guard against no width/height when parsing an image message with a source part (AND-1501)
   
 ### Bug Fixes
   * Fixed crash when no size is found for an image when sending an image message (AND-1485)